+++ PeckettIIR.h
@@ -6,4 +6,4 @@
 #include "config.h"
 
 void PeckettIIRSetup();
-void PeckettIIR(uint16_t val, DigitalPin<BEAT_PIN> beat_pin);
+void PeckettIIR(uint8_t val, DigitalPin<BEAT_PIN> beat_pin);
