+++ PeckettIIRFixedPoint.h
@@ -6,4 +6,4 @@
 #include "config.h"
 
 void PeckettIIRFixedPointSetup();
-void PeckettIIRFixedPoint(uint16_t val, DigitalPin<BEAT_PIN> beat_pin);
+void PeckettIIRFixedPoint(uint8_t val, DigitalPin<BEAT_PIN> beat_pin);
